LayoutTests/fast/url/invalid-url-pattern-context-crash.html

<html>
<head>
    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>
</head>
<body>
    <iframe id=testFrame src="/"></iframe>
    <script>
        promise_test(async test => {
            const contentWindow = testFrame.contentWindow;
            const FrameURLPattern = contentWindow.URLPattern;

            testFrame.remove();

            const result = (new FrameURLPattern()).exec();
            assert_equals(result, null);
        }, "URLPattern with stopped document");
    </script>
    <p>This test passes if WebKit does not crash.</p>
</body>
</html>
